01_scraping_manganews/requirements.txt
scrapy>=2.11
psycopg2-binary
great_expectations
pandas>=1.5
python-dotenv>=1.0
